Move plotting functions to module
